Move common tests to FetchByIdentifierAndYearHandlerTest
