+++ src/test/java/no/sikt/nva/pubchannels/handler/fetch/FetchSerialPublicationByIdentifierAndYearHandlerTest.java
@@ -1,0 +1,8 @@
+package no.sikt.nva.pubchannels.handler.fetch;
+
+// Common tests for FetchSerialPublicationByIdentifierAndYearHandler,
+// FetchSeriesByIdentifierAndYearHandler and FetchJournalByIdentifierAndYearHandler
+// are tested here
+public abstract class FetchSerialPublicationByIdentifierAndYearHandlerTest extends FetchByIdentifierAndYearHandlerTest {
+
+}
